deepen: delete dead StreamingParser + ParserMetricsCollector orbital
The Response Parser module (lib/ai/parsers/) was a four-file cluster
where two files had zero non-test callers and the types file existed
mostly to serve them.

Deletion test on each:

- `lib/ai/parsers/streaming.ts` (370 LOC). `StreamingParser`,
  `createStreamingParser` — zero non-test callers. The post-Phase-3
  pipeline reaches OpenRouter via `enhanced-claude-client.ts`'s batch
  path, which uses `StreamHandler` from `lib/ai/streaming/` (a
  different module). The parsers-package `streaming.ts` was the
  pre-Phase-3 SSE shim and has been dormant since `Phase 3:
  Simplified` was banner-commented into the surviving files.
- `lib/ai/parsers/metrics.ts` (234 LOC). `ParserMetricsCollector` +
  the `parserMetrics` singleton — zero non-test callers. Production
  metrics flow through `lib/monitoring/`'s `monitoring.recordMetric`
  (a separate interface that summarize.ts and
  summarize-with-validation.ts use). The parsers-package collector
  was telemetry infrastructure for a parser dashboard that was never
  wired up.
- `lib/ai/parsers/types.ts` (89 LOC). Of the seven exported types,
  only `ParserMetrics` had a real caller (response-parser.ts:567).
  The rest (`ExtractedJSON`, `ExtractionOptions`, `ValidationResult`,
  `StreamingParserState`, `StreamingOptions`, `StreamingProgressEvent`,
  `StreamingEventListener`) were used only by streaming.ts and
  metrics.ts, both deleted.

After the deletions, simple-parser.ts and response-parser.ts are the
only modules at the parsers seam — `parseResponse` (response-parser)
and `parseJSONResponse` (simple-parser) are the two-method interface
the Summarize module crosses. Both call sites
(`lib/ai/summarize.ts:14` via the barrel, `lib/sec-edgar/index.ts:12`
re-export) keep working unchanged.

`ParserMetrics` moves inline into response-parser.ts where its sole
caller already lives, mirroring ADR-0002's precedent for one-caller
pure types extracted-for-testability.

Dependency category: in-process (per process/Deepening/deepening.md
§1). No adapter required; deletion concentrates nothing because
nothing was there.

## Tests

- `__tests__/ai/parsers/` (144 tests across simple-parser,
  simple-parser-bracket-repair, response-parser-normalization,
  x-sentiment-validator-citations) — all pass.
- No streaming/metrics-specific tests existed, so none are deleted.
- Pre-existing 6 failures in
  `__tests__/ai/integration/parsing-integration.test.ts` confirmed
  present on `main` at HEAD; unrelated to this change.

## ADRs respected

- ADR-0006 — financial-content-gate.ts is untouched.
- ADR-0003 — extractor-registry stays absent.
- ADR-0001 — filing-type-registry inline pattern is unchanged.
